First Commit To Refactor
