--- notes.txt
@@ -1,5 +1,0 @@
-Level one Drawing complete -
-!todo---
-
-Code simpleGroup Movement
-Code simple win
