+++ src/Blocks/Block.js
@@ -1,11 +1,11 @@
-import { Vector } from "../Vectors.js";
+import { Vector } from "../Math/Vector.js";
 
-export class Block {
-  constructor(x, y, name, type) {
-    this.position = new Vector(x, y);
-    this.name = name;
-    this.Width = 10;
-    this.type = type;
-    
-  }
+
+export default class Block {
+    constructor(x, y, name, type) {
+        this.position = new Vector(x,y);
+        this.name = name;
+        this.type = type;
+        this.shove = new Vector(0,0);
+    }
 }
