+++ src/Blocks/BlocksCache.js
@@ -1,0 +1,4 @@
+export let spriteBlocks = [];
+export let nounBlocks=[];
+export let verbBlocks=[];
+export let operatorBlocks=[];
