+++ src/GameFiles/Rules.js
@@ -1,0 +1,8 @@
+import { objects } from "./blockDraw.js";
+import { GameInfo } from "./gameInfo.js";
+
+export let Rules = {};
+
+export function setRules() {
+
+}
